SyntaxList<TNode>: Rename file and move nested types
